Refactor all models from models.py into individual files
